pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "recap"
version = "0.2.0"
description = "Meeting recording → transcription → analysis → Obsidian vault notes"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "ruamel.yaml>=0.18",
]

[project.optional-dependencies]
daemon = [
    "aiohttp>=3.9",
    "pystray>=0.19",
    "Pillow>=10.0",
    "keyring>=25.0",
    "plyer>=2.1",
    "authlib>=1.3",
    "pywin32>=306",
    "PyAudioWPatch>=0.2.12",
    "pyflac>=2.2",
    "uiautomation>=2.0",
]
ml = [
    "nemo_toolkit[asr]>=2.0",
    "torch>=2.1",
    "pyarrow<21",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-aiohttp>=1.0",
    "aiohttp>=3.9",
    "pytest-cov>=5.0",
]

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cu126"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=recap --cov-fail-under=70 --cov-report=term-missing -m 'not integration'"
markers = [
    "integration: tests that import real ML/system libraries; slow; opt-in via -m integration",
]
